backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
oss2==2.18.5

# 数据库相关 - PostgreSQL
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# 邮件验证
email-validator==2.1.0

# 文件处理
aiofiles==23.2.1
httpx==0.27.0
beautifulsoup4==4.12.3